terminology: grapheme cluster -> grapheme
Per Unicode definition,
Grapheme: the fundamental unit in written language
Grapheme Cluster: A higher level of abstraction for character. Include
graphemes, and some other code point sequences (e.g. \r\n).

To be precise, the hints for what is written should be what the users
perceive as a single character.